CC(Backend)/RunFix/requirements.txt
Flask
tensorflow
numpy
google-cloud-storage